requirements.txt
argparse
openstacksdk
paramiko
